Implements multiple trackers in the iOS plugin
Notable changes are the addition of #setMultipleTrackingIds(), as well as the refactoring of #get().

This also upgrades iOS Google Analytics lib to v3.17
The new version also fixes the following issue that occurs when using the #get() method for keys for which no value could be fetched:

"-[GAITrackerModel valueForKey:]: attempt to retrieve a value for a nil key"

The #get() parameter will now return an empty string instead.