Feature/#64 multiple trackers (#73) fixes #64
* Extends the cordova interface to support #setMultipleTrackingIds()

This new method accepts a string array of tracking ids to be used.

* Implements multiple trackers in the Android plugin

A List of Tracker objects is used. All getter/setter methods are applied to all trackers.

Fix NPE and json <null> values for #get() with multiple trackers

Minor doc fixes

Fix trackingIds parameter

* Implements multiple trackers in the iOS plugin

Notable changes are the addition of #setMultipleTrackingIds(), as well as the refactoring of #get().

This also upgrades iOS Google Analytics lib to v3.17
The new version also fixes the following issue that occurs when using the #get() method for keys for which no value could be fetched:

"-[GAITrackerModel valueForKey:]: attempt to retrieve a value for a nil key"

The #get() parameter will now return an empty string instead.

* Fix NPE for json null values in #get() with multiple trackers

Also update documentation and example

* Minor docs improvement

* Change iOS #get() results for consistency

For multiple trackers, Android and iOS now both return <null> JSON
values for fields for which no results could be fetched, e.g.,

navigator.analytics.get('&dummy', function(result) { ...

will return:

"[{"UA-XXXXX-1": null}, {"UA-XXXXX-2": null}]"

* Fix raw types and list clear in Android plugin

The trackers iteration I added in #close() was (I believe) pointless.

* Minor iOS NSMutableArray optimization

Create the NSMutableArray only once and clear it in case new trackers are assigned or #close() is called.

* Fix iOS "tracker(s) not initialized" error check

Check if the NSMutableArray is nil or empty and print the error if necessary.